edr_downloader/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-decouple==3.8
reportlab==4.0.7
